wip: more modularization
