--- core/src/main/java/com/dehidehidehi/twitchtagcarousel/service/twitch/impl/HelixClient.java
@@ -1,6 +1,0 @@
-package com.dehidehidehi.twitchtagcarousel.service.twitch.impl;
-import jakarta.inject.Qualifier;
-
-@Qualifier
-public @interface HelixClient {
-}
