--- core/src/main/java/com/dehidehidehi/twitchtagcarousel/ui/CarouselUi.java
@@ -1,8 +1,0 @@
-package com.dehidehidehi.twitchtagcarousel.ui;
-import com.dehidehidehi.twitchtagcarousel.service.twitch.TwitchService;
-
-public interface CarouselUi {
-    
-    void start(TwitchService twitchService);
-    
-}
