--- core/src/main/java/com/dehidehidehi/twitchtagcarousel/ui/swing/annotation/SwingCarouselUi.java
@@ -1,7 +1,0 @@
-package com.dehidehidehi.twitchtagcarousel.ui.swing.annotation;
-
-import jakarta.inject.Qualifier;
-
-@Qualifier
-public @interface SwingCarouselUi {
-}
