+++ main/src/main/java/com/dehidehidehi/twitchtagcarousel/producer/UiProducer.java
@@ -1,0 +1,14 @@
+package com.dehidehidehi.twitchtagcarousel.producer;
+
+import com.dehidehidehi.twitchtagcarousel.CarouselUi;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.inject.Produces;
+
+@ApplicationScoped
+class UiProducer {
+	
+	@Produces
+	CarouselUi produceCarouselUi() {
+		return new com.dehidehidehi.twitchtagcarousel.swing.SwingCarouselUi();
+	}
+}
