+++ main/src/main/resources/WEB-INF/html/token_redirect.html
