+++ tags/src/main/java/com/dehidehidehi/twitchtagcarousel/domain/TwitchTagBatch.java
