+++ tags/src/main/java/com/dehidehidehi/twitchtagcarousel/domain/TwitchTagEnum.java
