+++ tags/src/main/java/com/dehidehidehi/twitchtagcarousel/error/TwitchChannelIdException.java
