+++ tags/src/main/java/com/dehidehidehi/twitchtagcarousel/error/TwitchTagUpdateException.java
