+++ ui/src/main/java/com/dehidehidehi/twitchtagcarousel/CarouselUi.java
@@ -1,0 +1,9 @@
+package com.dehidehidehi.twitchtagcarousel;
+
+import com.dehidehidehi.twitchtagcarousel.service.TwitchTagService;
+
+public interface CarouselUi {
+    
+    void start(TwitchTagService twitchTagService);
+    
+}
