+++ ui/src/main/java/com/dehidehidehi/twitchtagcarousel/swing/button/ExitButton.java
@@ -1,0 +1,10 @@
+package com.dehidehidehi.twitchtagcarousel.swing.button;
+import javax.swing.*;
+
+public class ExitButton extends JButton {
+
+	public ExitButton() {
+		super("Exit");
+		this.addActionListener(e -> System.exit(0));
+	}
+}
